requirements.txt
openai-whisper
torch
transformers
sentencepiece
accelerate
srt
ffmpeg-python
